Backend/requirements.txt
fastapi
uvicorn
spacy
networkx
sentence-transformers
python-multipart
pydantic-settings
python-dotenv
